requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-extensions==2.1.4
django-filter==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.9.1
ipython==7.2.0
Pillow==5.4.1
psycopg2-binary==2.7.6.1
